Bug 1466936 - Distribute via spark-packages.org
This PR changes deployment for both the python and Scala packages.
In addition to deploying to Mozilla's S3 Maven repo, we deploy to
spark-packages.org so that both Scala and python bindings are available
via invoking Spark with the --packages option set.

We also move all CI to CircleCI 2.0, update some dependency versions, etc.